エンドポイント url の変更
