+++ api/main.py
@@ -2,7 +2,7 @@ import json
 from fastapi import FastAPI
 from fastapi.responses import FileResponse
 
-from routers import badge, guild, ranking, studytime, user
+from routers_v1 import badge, guild, ranking, studytime, user
 
 from utilities.db_access import DBAccess as db
 
@@ -14,8 +14,6 @@ with open('/usr/srb3/source/conf.json', 'r', encoding='utf8') as f:
     db.password(conf['password'])
     db.database(conf['db'])
 
-print(db.select('mst_roles', ['id', 'name'], 'name=%s', ('灰筆',)))
-print(db.insert('test', ['number', 'string'], ['999', 'abc']))
 
 app = FastAPI(
     title = '自習室API',
