+++ api/routers_v1/ranking.py
@@ -2,10 +2,10 @@ from fastapi import APIRouter
 
 router = APIRouter()
 
-@router.get('/ranking/global', tags=['ranking'])
+@router.get('/v1/ranking/global', tags=['ranking'])
 async def getRankingGlobal():
     return {'ranking': []}
 
-@router.get('/ranking/local/{guild_id}', tags=['ranking'])
+@router.get('/v1/ranking/local/{guild_id}', tags=['ranking'])
 async def getLocalRanking(guild_id):
     return {'ranking': []}
