+++ api/routers_v1/studytime.py
@@ -2,10 +2,10 @@ from fastapi import APIRouter
 
 router = APIRouter()
 
-@router.get('/studytime/{user_id}', tags=['studytime'])
+@router.get('/v1/studytime/{user_id}', tags=['studytime'])
 async def getStudytime(user_id):
     return {'time': []}
 
-@router.post('/studytime/{user_id}', tags=['studytime'])
+@router.post('/v1/studytime/{user_id}', tags=['studytime'])
 async def postStudytime(user_id):
     return
